requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
install==1.3.5
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
